egm/requirements.txt
Django==1.11
psycopg2==2.7.4
django-import-export
